Adding serialization decorations where missing.
